Fix theme errors on Linux, refactor themes.
Font and org theme now use the display specification argument on their
face specs to apply only to graphical displays, removing the need for
so many daemon and display checks for consumers that load/enable the
theme. The server-make-frame hooks have been offloaded from the init
file into the theme files, and the themes no longer fail when fonts
are not available, as is the case in terminal environments.

Also fixed ACM not playing nice with theme switching.